Added smarter date handling to the JpaUpdateBuilder (used to send updates to sqlite clients)
If a JPA field is marked by @Temporal(TemporalType.Date), the column in the
sqlite database will be created as a 'TEXT' column and the date stored as
a date string like '2011-01-31'.

All other temporal types will be serialized as unix time (milliseconds
since jan 1 1970 utc).

This ensures that the date can be safely read from a client browser in a
different timezone than UTC.

(The previous behavior was to serialize all Dates as posix time stamps at
midnight, which means that Jan 15 2011 0:00 UTC could be read from a client as
Jan 14 2011 23:00 in another timezone-- BAD)

Also removed bulk executer api from sync module, which has been moved to
rebar-sql