Augments SqlQuery DSL with selectDistinct() and a methods supporting MySQL-specific optimization
To improve performance, it becomes necessary to really optimize against the specific databases.
So we add an isMySQL() method to SqlDialect interface so that some of the DAO code can break
out and do really MySQL-specific stuff that doesn't really have an analogue in sqlite. For example,
we want to avoid executing a query twice to get the page and then the query count. With MySQL
we can use the SQL_CALC_FOUND_ROWS keyword and then execute SELECT FOUND_ROWS().

To support this kind of use case, there is also now an addKeyword() method on the SqlQuery
class which can be used to add arbitrary keywords to the select statement.