Creates new async module to faciliate ridicuously asynchronous programming
As soon as you begin chaining a few asynchronous process together in Java/GWT,
it gets ugly quickly. We've had issues with exceptions leaking through and
hard-to-find typos that where the callback thread is lost.

The idea here is to borrow some ideas from functional programming and
decompose our complex processes into simple functions that can be composed
(chained) rather than needing 16 levels of callbacks.

The api will likely change significantly as we experiment with this
in ActivityInfo